requirements.txt
numpy==1.26.4
torch==2.3.0
torchvision==0.18.0
transformers==4.42.3
accelerate==0.31.0
diffusers[torch]==0.29.2
Pillow==10.3.0
filelock==3.15.4
xformers==0.0.26.post1
tqdm==4.66.4
nvidia-pytriton==0.5.8
python-dotenv==1.0.1
scipy==1.14.0
pybase64==1.3.2
https://github.com/chengzeyi/stable-fast/releases/download/v1.0.5/stable_fast-1.0.5+torch230cu121-cp311-cp311-manylinux2014_x86_64.whl
wheel
openai==0.27.2
matplotlib
imageio[ffmpeg]
einops 
auto-gptq
tiktoken
transformers_stream_generator
text-generation  # TGI and TGI Gaudi
opencv-python-headless
pycocotools